requirements.txt
openai>=1.12.0
PyYAML>=6.0
requests>=2.31.0
loguru>=0.7.2
tenacity>=8.2.3

